renamed to users from user ont table connection added test coverage
