+++ database/seeds/01-users.js
@@ -3,11 +3,11 @@ const bcrypt = require("bcryptjs");
 exports.seed = function (knex) {
   
     // Deletes ALL existing entries
-    return knex('user')
+    return knex('users')
     .del()
       .then(function () {
         // Inserts seed entries
-        return knex('user').insert([
+        return knex('users').insert([
           {
             username: 'BOT101',
             password: bcrypt.hashSync("Password350", 12)
